Rename 'unknown' state in preparation for adding 'unknown' alerts
"Unknown" is an overloaded term: Icinga uses it to refer to unclassified
alerts which are not "critical" or "warning", and this codebase was
using it to refer to the situation when Icinga is unavailable.

Rename the second usage so that we can use "unknown" for Icinga alert
classification in later commits.