Add 'unknown' alerts to the Blinken dashboard
Add a count of the 'unknown' alerts next to the 'critical' and 'warning'
alerts.

Add a pink background if there are only 'unknown' alerts to match the
Icinga dashboard.

This will alert 2ndline to any alerts in the 'unknown' category. These
often indicate that Graphite data is not being reported correctly or
that alerts are configured incorrectly, which means that we might miss
real issues.