Add the CI environment so we're monitoring it!
The CI environment has lots of errors that are getting missed because
it's missing from the 2ndline dashboard.

Adding another environment caused the row to wrap badly because the
col-xs-4 class was being applied to the content, instead of a separate
container. This is now fixed with some formatting to see the HTML
content more clearly + CSS tweaks to cope with the new DOM structure.