Update default config environment names
It used to be that there were separate environments running in AWS,
but now Production and Staging as environments span both Carrenza and
AWS infrastructure, and there are important things that occur between
them.

Therefore, refer to the Icinga instances by the environment first,
then specify the hosting provider if there's more than one for that
environment.